Descriptor parser support (#72975)
Add DescriptorMarker support from linker, this allows the NativeAOT compiler to read embedded xml files in an assembly called ILLink.Descriptors.xml. Then this file will be processed and used to add members to the compilation as roots. Is a very similar functionality to rd.xml with the difference that this format is Cecil based and embedded in the assemblies. Several of the libraries already insert this xml into the assembly to be trimming safe. This PR allows reading those XML files.
Add node to represent loading an assembly embedded file and checking for its dependencies
Verify for embedded files every time a module constructor is called
Support for events and properties in the XML reading
Add nodes to the dependency graph instead of marking logic
Add tests for embedded descriptor file in the NativeAOT smoke tests infra