Propagate BBF_HAS_MDARRAYREF to new blocks when copying IR (#73351)
Several block flags are used to indicate the presence of certain IR nodes
to avoid visiting blocks unnecessarily. To make this work, though, these
flags must be added to new blocks when code is copied from one block to
another. Add `BBF_HAS_MDARRAYREF` to the set of flags that need such
copying, and define and use `BBF_COPY_PROPAGATE` as the set of such flags.

Add a Fuzzlyn-derived unit test.

Fixes #72768