Fix ThreadAbort on Unix and reenable the related tests (#73399)
* Fix ThreadAbort on Unix and reenable the related tests

When the HandleSuspensionForInterruptedThread invokes Thread::HandleThreadAbort
and the thread abort exception is thrown from there, runtime fails with
unhandled C++ exception because there was a missing guard that catches the
PAL_SEHException and invokes DispatchManagedException to propagate it further
through managed code.
This change fixes that and also addresses an issue with activation injection
in Thread::UserAbort in the case of infinite timeout (it would not inject the
activation in that case).

On Alpine Linux, libunwind cannot unwind correctly over signal frames. We have been
handling that for hardware exceptions, but with the new possibility of thread abort
exception being thrown from the activation handler, we need to apply the same treatment
to the activation signal too. We save a location of the windows style context in the
activation_injection_handler and a return address in this function for a call to the
actual registered handler invocation. Then in PAL_VirtualUnwind, when we detect that
we are trying to unwind from this location, we use the saved context instead of calling
libunwind.