More precise writebarrier for regions (#67389)
This introduces a lookup table for regions where we can find the current generation, the planned generation and a couple flags efficiently, i.e. by simple indexing into table.

The table has byte-sized elements where the low 2 bits represent the current generation and the high 2 bits represent the planned generation. The 4 bits in between are available for flags, 2 are used so far, one to show that a region was swept in the plan phase, and one to show that a region was demoted.

The table is used in the mark phase, in the mark_through_cards_helper (which is also called in the relocate_phase), and in the write barriers (for now only the most frequently used ones, Array.Copy has its own way of setting cards that I haven't fixed).

I have changed the write barrier to only set single bits for the case where a pointer to younger generation is stored into an object in an older generation. This costs an interlocked operation in the case the bit is not already set. In the test cases we looked at this is more than compensated by lower cost in card marking. For cases where the cost of the interlocked instruction is too high, an escape hatch has been provided the form of the COMPLUS_GCWriteBarrier setting which can be used to select another flavor of write barrier - either a more precise write barrier for regions, or an imprecise write barrier. Both set whole bytes and thus need no interlocked operations.

The write barrier uses a global ephemeral range to decide quickly whether a stored pointer can be in an ephemeral generation at all. For now, this global ephemeral range is simply set to the address range of the whole heap, to avoid having to update the write barrier on the fly. A slightly more sophisticated approach that takes into account low gen 2 regions is present in the code, but disabled for now.