CI: Add support in yml for detecting darc dependency changes in `eng/Version.Details.xml` (#73435)
## Issue

Currently, whenever a darc flow PR is opened, we cannot trigger different jobs based on which
dependency got updated. This is needed because in some cases, like for ILLinker, Emscripten
updates we want to trigger all the wasm jobs. But that might not be needed for some of the other
dependencies.

- This causes failures to slip through to `main`, which gets discovered after the fact as other PR
failures, or rolling build failures, creating extra work for developers.

## Solution

- This PR identifies the changed dependencies, and emits one azure variable per dependency,
which can then be used in conditions in the yml .
- The changed dependency can be checked as `dependencies.evaluate_paths.outputs['DarcDependenciesChanged.System_CommandLine']`

- Included this in the `evaluate_paths` job itself, instead of creating another job

- Also, use this for wasm jobs