Make PublishAot properly incremental (#73989)
`NativeLibrary` is an input to `LinkNative` target. Since these are not fully qualified paths (we rely on link.exe doing a thing to find them) the target always re-runs. It should not.

The ideal fix would be to make this paths fully qualified, but I don't see an easy way to do that. SDK libraries don't really change in meaningful ways.