Add jit stress random (#73859)
* Add support for JitStress with random stress number

There are two new pipelines:
1. runtime-coreclr jitstress-random (https://dev.azure.com/dnceng/public/_build/results?buildId=1950662&view=results)
2. runtime-coreclr libraries-jitstress-random (https://dev.azure.com/dnceng/public/_build/results?buildId=1950664&view=results)

The first runs coreclr tests, the second runs libraries tests.
Each individual job gets its own JitStress random number. The
random number is set in src/tests/Common/testenvironment.proj
such that its value is visible in the log file. Currently, the
random number is the milliseconds value of the current time,
so it is between 0 and 999.

The pipelines are each configured to run once per week, on weekends.

* Remove unused jitstress_random_*