Fix Linux armv6 build on runtime-community.yml (#74757)
Add missing HOST_ARM define to armv6 in configurecompiler.cmake, it was missed in https://github.com/dotnet/runtime/pull/73530, before that we defined `HOST_ARM` for armv6 in pal.h too because of `defined(__arm__)`

We also only need the `clr.iltools+clr.packages` subset, not all of CoreCLR.

Also fix some dependencies that weren't working for FreeBSD since we missed the local variables. We don't need installer subset conditions there but we need to trigger on rolling builds.