SPMI: Add and utilize number of contexts with diffs  (#76011)
Previously, if there was any diff in a collection, that collection would
be shown in all tables (i.e. Overall, FullOpts, MinOpts). The main
reason was that we only had "has diffs" information on a per-collection
basis, not for each of the categories. This changes SPMI to communicate
back for each category how many contexts had diffs in them, and uses
this to hide tables/rows without any diffs, and to show this information
under details.