Use IndexOfAny{Except}InRange in RegexCompiler / source generator (#76859)
* Use IndexOfAny{Except}InRange in RegexCompiler / source generator

This augments our existing use of IndexOf, IndexOfAny, and IndexOfAnyExcept to also support IndexOfAnyInRange and IndexOfAnyExceptInRange.  That means, for example, we can now efficiently find the start of a pattern like `[0-9]{5}`, via a vectorized search, whereas previously it'd require iterating character by character in a scalar loop.

As part of this, I changed some tuples to instead be named structs.  They were becoming unwieldy, and we expect we'll be adding even more here as additional IndexOf variants become available.

* Address PR feedback

And add a bit more test coverage