SPMI: Extend list of unrecorded environment variables (#77504)
We are currently hitting issues when there are diffs in the FileCheck tests because they are recorded with JitStdOutFile and the SPMI driver does not override this correctly. Add this and a lot of other environment variables to the list of variables we do not record.

This big list of string compares may look somewhat scary but we only query the config once during startup of the JIT, so it isn't a problem. I was not able to measure anything significant when collecting a simple "Hello, world!" C# app.